Implement the entire GLFW3 API
This API is generated by pruning everything from the glfw3.h file. It's currently untested and the three vulkan calls I'm not so sure about.

The problem with the vulkan calls is that they use structures that are not defined, and I do not know if they are pointers.

For now, I've made them pointers, since the docs say that you can pass in NULL.

One thing that should be done is to rename all outputs (automatically of course) so that they are of the form glfw:Name, but it's not that much of a problem.

Thanks to lexi-lambda and notjack from the racket IRC for the macro help!